+++ prism-agent/service/server/src/test/scala/io/iohk/atala/issue/controller/IssueControllerTestTools.scala
@@ -15,7 +15,11 @@ import io.iohk.atala.iam.authentication.DefaultEntityAuthenticator
 import io.iohk.atala.issue.controller.http.{CreateIssueCredentialRecordRequest, IssueCredentialRecord, IssueCredentialRecordPage}
 import io.iohk.atala.pollux.anoncreds.AnoncredLinkSecretWithId
 import io.iohk.atala.pollux.core.model.CredentialFormat
-import io.iohk.atala.pollux.core.repository.{CredentialDefinitionRepositoryInMemory, CredentialRepositoryInMemory}
+import io.iohk.atala.pollux.core.repository.{
+  CredentialDefinitionRepositoryInMemory,
+  CredentialRepositoryInMemory,
+  CredentialStatusListRepositoryInMemory
+}
 import io.iohk.atala.pollux.core.service.*
 import io.iohk.atala.pollux.vc.jwt.*
 import io.iohk.atala.shared.models.{WalletAccessContext, WalletId}
@@ -32,7 +36,6 @@ import zio.config.{ReadError, read}
 import zio.json.ast.Json
 import zio.json.ast.Json.*
 import zio.test.*
-
 import java.util.UUID
 
 trait IssueControllerTestTools extends PostgresTestContainerSupport {
@@ -78,6 +81,7 @@ trait IssueControllerTestTools extends PostgresTestContainerSupport {
     didResolverLayer >+>
     ResourceURIDereferencerImpl.layer >+>
     CredentialRepositoryInMemory.layer >+>
+    CredentialStatusListRepositoryInMemory.layer >+>
     ZLayer.succeed(AnoncredLinkSecretWithId("Unused Linked Secret ID")) >+>
     MockDIDService.empty >+>
     MockManagedDIDService.empty >+>